backend/requirements.txt
fastapi==0.92.0
tortoise-orm==0.19.3
uvicorn==0.20.0
httpx==0.23.3